Adding docs for sw-precaching
