+++ jsdoc.conf
@@ -3,7 +3,7 @@
     "include": [
       "./packages"
     ],
-    "excludePattern": "((^|\\/|\\\\)build(\\/|\\\\)|(^|\\/|\\\\)test(\\/|\\\\))"
+    "excludePattern": "((^|\\/|\\\\)build(\\/|\\\\)|(^|\\/|\\\\)test(\\/|\\\\)|(^|\\/|\\\\)node_modules(\\/|\\\\))"
   },
   "plugins": [
     "node_modules/jsdoc-strip-async-await",
