+++ packages/sw-precaching/src/lib/controllers/unrevisioned-cache-manager.js
@@ -1,0 +1,93 @@
+import ErrorFactory from '../error-factory';
+import BaseCacheManager from './base-cache-manager';
+import RequestPrecacheEntry from '../models/precache-entries/request-entry';
+import {defaultUnrevisionedCacheName} from '../constants';
+
+/**
+ * This class extends a lot of the internal methods from BaseCacheManager
+ * to manage caching of unrevisioned assets.
+ *
+ * @memberof module:sw-precaching
+ * @extends {module:sw-precaching.BaseCacheManager}
+ */
+class UnrevisionedCacheManager extends BaseCacheManager {
+  /**
+   * Constructor for UnreivisionedCacheManager
+   */
+  constructor() {
+    super(defaultUnrevisionedCacheName);
+  }
+
+  /**
+   * This method will add the entries to the install list.
+   * This will manage duplicate entries and perform the caching during
+   * the install step.
+   *
+   * @example
+   *
+   * revisionedManager.cache([
+   *   '/styles/hello.css',
+   *   new Request('/images/logo.png', {
+   *     // Custom Request Options.
+   *   })
+   * ]);
+   *
+   * @param {Array<String|Request>} rawEntries A raw entry that can be
+   * parsed into a BaseCacheEntry.
+   */
+  cache(rawEntries) {
+    // This method is here to provide useful docs.
+    super(rawEntries);
+  }
+
+  /**
+   * This method ensures that the file entry in the maniest is valid and
+   * can be parsed as a BaseCacheEntry.
+   *
+   * @private
+   * @abstract
+   * @param {String | Request} input Either a URL string or a Request.
+   * @return {BaseCacheEntry} Returns a parsed version of the file entry.
+   */
+  _parseEntry(input) {
+    if (typeof input === 'undefined' || input === null) {
+      throw ErrorFactory.createError('invalid-unrevisioned-entry',
+        new Error('Invalid file entry: ' + JSON.stringify(input)));
+    }
+
+    return new RequestPrecacheEntry(input);
+  }
+
+  /**
+   * @private
+   * @param {BaseCacheEntry} newEntry The entry that is to be cached.
+   * @param {BaseCacheEntry} previousEntry The entry that is currently cached.
+   */
+  _onDuplicateInstallEntryFound(newEntry, previousEntry) {
+    // NOOP. Just ignore duplicate entries.
+  }
+
+
+  /**
+   * We always want to override previously cached versions.
+   * @private
+   * @param {BaseCacheEntry} precacheEntry A file entry with `path` and
+   * `revision` parameters.
+   * @return {Promise<Boolean>} Returns true is the fileEntry is already
+   * cached, false otherwise.
+   */
+  async _isAlreadyCached(precacheEntry) {
+    return false;
+  }
+
+  /**
+   * @private
+   * @param {BaseCacheEntry} precacheEntry A file entry with `path` and
+   * `revision` parameters.
+   */
+  _onEntryCached(precacheEntry) {
+    // NOOP
+  }
+}
+
+export default UnrevisionedCacheManager;
