+++ packages/sw-precaching/src/lib/precache-manager.js
@@ -14,10 +14,28 @@
 */
 
 import assert from '../../../../lib/assert';
-import RevisionedCacheManager from './controllers/revisioned-manager.js';
-import UnrevisionedCacheManager from './controllers/unrevisioned-manager.js';
+import RevisionedCacheManager from './controllers/revisioned-cache-manager.js';
+import UnrevisionedCacheManager from
+  './controllers/unrevisioned-cache-manager.js';
 
+/**
+ * The PrecacheManager is the top level API you are likely to use with
+ * the sw-precaching module.
+ *
+ * This class will set up the install listener and orchestrate the caching
+ * of assets via the [RevisionedCacheManager]{@link
+ *  module:sw-precaching.RevisionedCacheManager} and
+ * [UnrevisionedCacheManager]{@link
+ *  module:sw-precaching.UnrevisionedCacheManager}.
+ *
+ * @memberof module:sw-precaching
+ */
 class PrecacheManager {
+  /**
+   * Creating a PrecacheManager will add an install and activate event listener
+   * to your service worker. This allows the manager to cache assets and
+   * tidy up the no longer required assets.
+   */
   constructor() {
     this._eventsRegistered = false;
     this._revisionedManager = new RevisionedCacheManager();
@@ -26,9 +44,7 @@ class PrecacheManager {
   }
 
   /**
-   * This method registers the service worker events. This should only
-   * be called once in the constructor and will do nothing if called
-   * multiple times.
+   * @private
    */
   _registerEvents() {
     if (this._eventsRegistered) {
@@ -75,16 +91,68 @@ class PrecacheManager {
     });
   }
 
+  /**
+   * To cache revisioned assets (i.e. urls / assets that you have a revision
+   * for) can be efficiently cached and updated with this method.
+   * @param {Object} input
+   * @param {Array<String|Object>} input.revisionedFiles An array of URL strings
+   * , which should have revisioning in the file name (i.e. hello.1234.css)) or
+   * an object with a `url` and `revision` parameter.
+   *
+   * @example
+   * precacheManager.cacheRevisioned({
+   *   revisionedFiles: [
+   *     // Revision is in the file name
+   *     '/styles/main.1234.css',
+   *
+   *     // Object of url and revision can be used as well
+   *     {
+   *       url: '/',
+   *       revision: '1234'
+   *     },
+   *   ]
+   * });
+   */
   cacheRevisioned({revisionedFiles} = {}) {
     assert.isInstance({revisionedFiles}, Array);
     this._revisionedManager.cache(revisionedFiles);
   }
 
+  /**
+   * To cache URLs or assets where you don't know the revisioning, should
+   * be cached with this method. This method will always cache these files
+   * on install, regardless of whether they are already cached or not.
+   * This ensures they are up-to-date after a new service worker install.
+   * @param {Object} input
+   * @param {Array<String|Request>} input.unrevisionedFiles An array of URL
+   * strings or a Request object, which allows you to define custom headers.
+   *
+   * @example
+   * precacheManager.cacheUnrevisioned({
+   *   unrevisionedFiles: [
+   *     // Normal URL string
+   *     '/example/',
+   *
+   *     // Request with headers
+   *     new Request(
+   *       '/user-info.json',
+   *       {
+   *         headers: {
+   *           'CustomHeader': 'Hello World.'
+   *         }
+   *       }
+   *     ),
+   *   ]
+   * });
+   */
   cacheUnrevisioned({unrevisionedFiles} = {}) {
     assert.isInstance({unrevisionedFiles}, Array);
     this._unrevisionedManager.cache(unrevisionedFiles);
   }
 
+  /**
+   * @private
+   */
   _close() {
     this._revisionedManager._close();
   }
